feat: ts interface 선언
