+++ src/types/diary.ts
@@ -5,9 +5,9 @@ export interface DiaryResponse {
   feeling_code: number
   img: string
   open: boolean
-  user: UserResponse
-  likes: LikesResponse[]
-  comments: CommentsResponse[]
+  user?: UserResponse
+  likes?: LikesResponse[]
+  comments?: CommentsResponse[]
   createdAt: string
   updatedAt: string
 }
@@ -16,6 +16,7 @@ export interface LikesResponse {
   id: number
   createdAt: string
   updatedAt: string
+  post?: NewPost
 }
 
 export interface UserResponse {
@@ -32,4 +33,15 @@ export interface CommentsResponse {
   comment: string
   user: UserResponse
   createdAt: string
+}
+
+export interface NewPost {
+  body: string
+  id: string
+  createdAt: string
+  updatedAt: string
+  feeling_code: number
+  img: string
+  open: boolean
+  title: string
 }
