+++ src/types/user.ts
@@ -1,4 +1,17 @@
-import { DiaryResponse, UserResponse } from './diary'
+import { DiaryResponse, LikesResponse, NewPost, UserResponse } from './diary'
+import { FollowParent, FollowingParent } from './follow'
+
+export interface NewUser {
+  id: number
+  email: string
+  username: string
+  profile_image: string
+  profile_message: string
+  post: NewPost[]
+  following: FollowingParent[]
+  follower: FollowParent[]
+  likes: LikesResponse[]
+}
 
 export interface UserProfile {
   id: number
@@ -10,7 +23,7 @@ export interface UserProfile {
   likes: UserLikes[]
   following: {
     id: number
-    follwer: UserFollow
+    follower: UserFollow
   }[]
   follower: {
     id: number
@@ -43,4 +56,5 @@ export interface UserFollow {
   username: string
   profile_image: string
   profile_message: string
+  email: string
 }
